+++ lib/src/scaffold_layout_builder.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 
 typedef AppBarAnimatedBuilder = Widget Function(
@@ -14,9 +16,9 @@ class ColorAnimated {
 
 class ColorBuilder {
   final Color start;
-  final Color end;
+  final Color? end;
 
-  const ColorBuilder(this.start, this.end);
+  const ColorBuilder(this.start, [this.end]);
 }
 
 class ScaffoldLayoutBuilder extends StatefulWidget {
@@ -24,28 +26,36 @@ class ScaffoldLayoutBuilder extends StatefulWidget {
     Key? key,
     required this.backgroundColorAppBar,
     required this.textColorAppBar,
-    this.duration = const Duration(milliseconds: 100),
+    this.duration = const Duration(milliseconds: 200),
     required this.appBarBuilder,
-    required this.body,
-    this.appBarHeight = 100,
+    this.appBarHeight,
+    this.child,
   }) : super(key: key);
 
   final Duration duration;
 
   /// Example:
   /// ```
+  /// // * Start Color
+  /// // * End Color
   /// ColorBuilder(Colors.transparent, Colors.white)
   /// ```
+  ///
+  /// if end color is null, then the end color will be the same as the start color
   final ColorBuilder backgroundColorAppBar;
 
   /// Example:
   /// ```
+  /// // * Start Color
+  /// // * End Color
   /// ColorBuilder(Colors.white, Colors.blue)
   /// ```
+  ///
+  /// if end color is null, then the end color will be the same as the start color
   final ColorBuilder textColorAppBar;
-  final Widget body;
+  final Widget? child;
   final AppBarAnimatedBuilder appBarBuilder;
-  final double appBarHeight;
+  final double? appBarHeight;
 
   @override
   _ScaffoldLayoutBuilderState createState() => _ScaffoldLayoutBuilderState();
@@ -55,9 +65,11 @@ class _ScaffoldLayoutBuilderState extends State<ScaffoldLayoutBuilder>
     with TickerProviderStateMixin {
   late AnimationController animationController;
   late Animation background, color;
+  late double appBarHeight;
 
   @override
   void initState() {
+    super.initState();
     animationController = AnimationController(
       vsync: this,
       duration: widget.duration,
@@ -66,15 +78,21 @@ class _ScaffoldLayoutBuilderState extends State<ScaffoldLayoutBuilder>
     // Set background color
     background = ColorTween(
       begin: widget.backgroundColorAppBar.start,
-      end: widget.backgroundColorAppBar.end,
+      end: widget.backgroundColorAppBar.end ??
+          widget.backgroundColorAppBar.start,
     ).animate(animationController);
 
     // Set text color
     color = ColorTween(
       begin: widget.textColorAppBar.start,
-      end: widget.textColorAppBar.end,
+      end: widget.textColorAppBar.end ?? widget.textColorAppBar.start,
     ).animate(animationController);
-    super.initState();
+
+    if (widget.appBarHeight == null) {
+      appBarHeight = Platform.isAndroid ? 90 : 100;
+    } else {
+      appBarHeight = widget.appBarHeight!;
+    }
   }
 
   @override
@@ -83,9 +101,9 @@ class _ScaffoldLayoutBuilderState extends State<ScaffoldLayoutBuilder>
       onNotification: _scrollNotification,
       child: Stack(
         children: [
-          widget.body,
-          Container(
-            height: widget.appBarHeight,
+          if (widget.child != null) widget.child!,
+          SizedBox(
+            height: appBarHeight,
             child: AnimatedBuilder(
               animation: animationController,
               builder: (context, _) {