+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+
+* Change parameter body to child.
+* Nullable parameter "end" on ColorBuilder.
+* Compatibility
+
 ## 0.0.2
 
 * New release